+++ devcode-collector/HISTORY.md
@@ -1,6 +1,10 @@
 # devcode-collector
-  * 정보수집을 위한 모듈
+
+### 2023.09.01
+* Log4j2 라이브러리 설정
+* Log4j2 MDC 기능을 이용한 requestId를 생성해주는 필터적용
+  - com.github.devcode.collector.config.LoggingMdcFilter.java
   
 ### 2023.08.08
-  * devcode-collector 모듈 추가
-  * Log4j2 라이브러리 설정
+* devcode-collector 모듈 추가
+